Module vcx::api::connection

Functions

vcx_connection_connect
Establishes connection between institution and its user

vcx_connection_create
-> Create a Connection object that provides a pairwise connection for an institution's user

vcx_connection_create_with_invite
-> Create a Connection object from the given invite_details that provides a pairwise connection.

vcx_connection_delete_connection
connection object -> Delete a Connection object and release its handle

vcx_connection_deserialize
Takes a json string representing a connection object and recreates an object matching the json

vcx_connection_get_state

vcx_connection_invite_details
Gets the current connection details

vcx_connection_release
Releases the connection object by de-allocating memory

vcx_connection_serialize
Takes the Connection object and returns a json string of all its attributes

vcx_connection_update_state
Checks for any state change in the connection and updates the the state attribute