Module vcx::api::credential

Functions

vcx_credential_create_with_msgid
Create a Credential object that requests and receives a credential for an institution

vcx_credential_create_with_offer
Create a Credential object that requests and receives a credential for an institution

vcx_credential_deserialize
Takes a json string representing an credential object and recreates an object matching the json

vcx_credential_get_offers
Queries agency for credential offers from the given connection.

vcx_credential_get_payment_info
Retrieves Payment Info from a Credential

vcx_credential_get_payment_txn
Retrieve the txn associated with paying for the credential

vcx_credential_get_state

vcx_credential_release
Releases the credential object by de-allocating memory

vcx_credential_send_request
Send a credential request to the connection, called after having received a credential offer

vcx_credential_serialize
Takes the credential object and returns a json string of all its attributes

vcx_credential_update_state
Checks for any state change in the credential and updates the the state attribute. If it detects a credential it will store the credential in the wallet and update the state.

vcx_get_credential
Retrieve information about a stored credential in user's wallet, including credential id and the credential itself.