Module vcx::api::credential_def

Functions

vcx_credentialdef_create
Create a new CredentialDef object that can create credential definitions on the ledger

vcx_credentialdef_deserialize
Takes a json string representing a credentialdef object and recreates an object matching the json

vcx_credentialdef_get_cred_def_id
Retrieves credential definition's id

vcx_credentialdef_get_payment_txn
Retrieve the txn associated with paying for the credential_def

vcx_credentialdef_release
Releases the credentialdef object by de-allocating memory

vcx_credentialdef_serialize
Takes the credentialdef object and returns a json string of all its attributes