Function vcx::api::disclosed_proof::vcx_disclosed_proof_generate_proof

#[no_mangle]

pub extern "C" fn vcx_disclosed_proof_generate_proof(
    command_handle: u32, 
    proof_handle: u32, 
    selected_credentials: *const c_char, 
    self_attested_attrs: *const c_char, 
    cb: Option<extern "C" fn(xcommand_handle: u32, err: u32)>
) -> u32

Takes the disclosed proof object and generates a proof from the selected credentials and self attested attributes

#Params command_handle: command handle to map callback to user context.

handle: Proof handle that was provided during creation. Used to identify the disclosed proof object

selected_credentials: a json string with a credential for each proof request attribute. List of possible credentials for each attribute is returned from vcx_disclosed_proof_retrieve_credentials

Examples selected_credential -> "{"req_attr_0":cred_info}" Where cred_info is returned from retrieve credentials

self_attested_attrs: a json string with attributes self attested by user

Examples self_attested_attrs -> "{"self_attested_attr_0":"attested_val"}"

cb: Callback that returns error status

#Returns Error code as a u32