Module vcx::api::utils

Structs

Config

UpdateAgentInfo

Functions

vcx_agent_provision_async
Provision an agent in the agency, populate configuration and wallet for this agent. NOTE: for synchronous call use vcx_provision_agent

vcx_agent_update_info
Update information on the agent (ie, comm method and type)

vcx_ledger_get_fees
Get ledger fees from the sovrin network

vcx_messages_download
Retrieve messages from the specified connection

vcx_messages_update_status
Update the status of messages from the specified connection

vcx_provision_agent
Provision an agent in the agency, populate configuration and wallet for this agent. NOTE: for asynchronous call use vcx_agent_provision_async

vcx_set_next_agency_response