Module vcx::api::wallet

Functions

indy_add_wallet_record⚠

indy_delete_wallet_record⚠

indy_get_wallet_record⚠

indy_update_wallet_record_value⚠

vcx_wallet_add_record
Adds a record to the wallet Assumes there is an open wallet. #Params

vcx_wallet_add_record_tags
Adds tags to a record. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_close_search
Close a search

vcx_wallet_create_payment_address
Add a payment address to the wallet

vcx_wallet_delete_record
Deletes an existing record. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_delete_record_tags
Deletes tags from a record. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_export
Exports opened wallet

vcx_wallet_get_record
Deletes an existing record. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_get_token_info
Get the total balance from all addresses contained in the configured wallet

vcx_wallet_import
Creates a new secure wallet and then imports its content according to fields provided in import_config Cannot be used if wallet is already opened (Especially if vcx_init has already been used).

vcx_wallet_open_search
Opens a storage search handle

vcx_wallet_search_next_records
Fetch next records for wallet search.

vcx_wallet_send_tokens
Send tokens to a specific address

vcx_wallet_update_record_tags
Updates the value of a record already in the wallet. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_update_record_value
Updates the value of a record already in the wallet. Assumes there is an open wallet and that a type and id pair already exists. #Params

vcx_wallet_validate_payment_address
Validates a Payment address